Set up the layout for the Statistics View.
